requirements.txt
torchvision
torch
opencv-python
matplotlib
pandas
numpy
imutils
Pillow
pytesseract